Create NumberGame
A numeric guessing game. The programmer chooses the length of the number to be guessed (NUM_DIGITS) and the user has (MAX_GUESSES) to find the number. Clues are given if digits are correct, correct but in the wrong place, or wrong. When all the guesses have been exhausted, the game ends and the user is offered another chance to play.

Programmer can change the length of the number being guessed by altering variable NUM_DIGITS and can change the number of guesses by altering variable MAX_GUESSES.